+++ RELEASE_NOTES.md
@@ -1,3 +1,3 @@
-#### 0.4.22
+#### 0.4.23
 
-- re-added restoration of cursor position to flow setup
+- Syncing now adds a trailing \n to text so Obsidian's own saves don't trigger a false external edit notification if a region's source is open in a different tab